requirements.txt
environs==10.3.0
requests==2.31
python-telegram-bot==13.15